+++ Engine/src/EngineAPI.cpp
@@ -18,18 +18,18 @@ EngineAPI::~EngineAPI()
 }
 
 std::optional<size_t> EngineAPI::getNumLegalMoves(
-    const std::string& FEN, int32_t depth) const
+    const std::string& FEN, uint8_t depth) const
 {
     assert(engine != nullptr);
     return engine->getNumLegalMoves(FEN, depth);
 }
 
-SearchResult hceEngine::EngineAPI::getBestMove(const std::string& FEN, int32_t depth) const
+SearchResult hceEngine::EngineAPI::getBestMove(const std::string& FEN, uint8_t depth) const
 {
     return engine->getBestMove(FEN, depth);
 }
 
-SearchResult EngineAPI::getBestMoveMiniMax(const std::string& FEN, int32_t depth) const
+SearchResult EngineAPI::getBestMoveMiniMax(const std::string& FEN, uint8_t depth) const
 {
     assert(engine != nullptr);
     return engine->getBestMoveMiniMax(FEN, depth);
